Beta/v4 photometric aug/Variation A_ prototype/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scipy==1.11.1
matplotlib==3.7.2
